Move to correct conftest.py
